Criminal_Face_Detection/backend/requirements.txt
Flask
Flask-CORS
Pillow
opencv-python  # Uncomment if OpenCV is needed
tensorflow     # Uncomment if TensorFlow is needed
